Smite
Okay, so I remember when I was 13 I made a change some thing to do with a boy/girl smite. Well being the feminist I am I wanted to change it to human instead of boy right? So my sister played and got her self a /smite and she was angry that it said what it says "has been a naughty, naughty boy" So it has been changed. I don't mind if you change this or not but if  you do you will make my sister play TF again 

- Charles (kane_creeper05)